requirements.txt
torch>=2.0.0
transformers>=4.30.0
pillow>=9.0.0
streamlit>=1.22.0
numpy>=1.21.0